Refactor modules from `core` into `xarray.computation` (#10132)
