+++ xarray/__init__.py
@@ -15,11 +15,7 @@ from xarray.backends.zarr import open_zarr
 from xarray.coding.cftime_offsets import cftime_range, date_range, date_range_like
 from xarray.coding.cftimeindex import CFTimeIndex
 from xarray.coding.frequencies import infer_freq
-from xarray.conventions import SerializationWarning, decode_cf
-from xarray.core.alignment import align, broadcast
-from xarray.core.combine import combine_by_coords, combine_nested
-from xarray.core.common import ALL_DIMS, full_like, ones_like, zeros_like
-from xarray.core.computation import (
+from xarray.computation.computation import (
     apply_ufunc,
     corr,
     cov,
@@ -29,6 +25,10 @@ from xarray.core.computation import (
     unify_chunks,
     where,
 )
+from xarray.conventions import SerializationWarning, decode_cf
+from xarray.core.alignment import align, broadcast
+from xarray.core.combine import combine_by_coords, combine_nested
+from xarray.core.common import ALL_DIMS, full_like, ones_like, zeros_like
 from xarray.core.concat import concat
 from xarray.core.coordinates import Coordinates
 from xarray.core.dataarray import DataArray
