+++ xarray/util/generate_ops.py
@@ -260,7 +260,8 @@ import operator
 from collections.abc import Callable
 from typing import TYPE_CHECKING, Any, overload
 
-from xarray.core import nputils, ops
+from xarray.core import nputils
+from xarray.computation import ops
 from xarray.core.types import (
     DaCompatible,
     DsCompatible,
